Move comparison helpers to String.Comparison.cs
